feat: Data 저장 코드 추가, 일부 코드 분활
